Moved test case.
