+++ src/main/scala/io/archivesunleashed/app/NERCombinedJson.scala
@@ -14,16 +14,16 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.archivesunleashed.matchbox
+package io.archivesunleashed.app
 
-import java.io.BufferedReader
-import java.io.BufferedWriter
-import java.io.InputStreamReader
-import java.io.OutputStreamWriter
+import java.io.{BufferedReader, BufferedWriter, InputStreamReader, OutputStreamWriter}
+
+import io.archivesunleashed.matchbox.NER3Classifier
+import io.archivesunleashed.util.JsonUtils
 import org.apache.hadoop.conf.Configuration
 import org.apache.hadoop.fs._
 import org.apache.spark.SparkContext
-import io.archivesunleashed.util.JsonUtils
+
 import scala.collection.mutable.MutableList
 import scala.util.Random
