+++ src/test/scala/io/archivesunleashed/app/ExtractPopularImagesTest.scala
@@ -14,20 +14,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.archivesunleashed.matchbox
+package io.archivesunleashed.app
 
 import com.google.common.io.Resources
+import io.archivesunleashed.RecordLoader
 import org.apache.spark.{SparkConf, SparkContext}
 import org.junit.runner.RunWith
 import org.scalatest.junit.JUnitRunner
 import org.scalatest.{BeforeAndAfter, FunSuite}
-import java.io.File
-import java.nio.file.{Files, Paths}
-
-import io.archivesunleashed.RecordLoader
-import io.archivesunleashed.app.ExtractPopularImages
-
-import scala.io.Source
 
 @RunWith(classOf[JUnitRunner])
 class ExtractPopularImagesTest extends FunSuite with BeforeAndAfter {
